Change how attach info filepath is stored and looked through for skeletons/masks
Rather than store a list of places that define attach info for a relic, store a list of haberdashery directory locations (local or in mods).

Fixes crash when trying to use an animated relic where the skeleton only exists inside a mod and the attach info only exists in local json.